requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
pydantic==1.10.2
sqlalchemy==1.4.41
pytest==7.4.0
httpx==0.23.0
mysql-connector-python==9.0.0